test: move the debug-assertions canary to a sibling test file
The canary landed inline in `src/lib.rs` and took `codecov/patch` red at
9/11 = 81.81%. The two missed lines were the interior of its own
`debug_assert!`:

    112 let evaluated = std::cell::Cell::new(false);   hit
    113 debug_assert!({                                hit
    114     evaluated.set(true);                       MISS
    115     true                                       MISS
    116 });

That is #1248's mechanism landing on the code written to demonstrate it:
under `ci-fast` the macro is `if false`, so llvm-cov emits regions for the
condition that no test can ever execute. Not a coverage gap — a
measurement artefact with no reachable fix at the line.

Moved to `src/lib_tests.rs` behind `#[cfg(test)] #[path] mod`, which is
this repo's documented home for test modules (CLAUDE.md) and where the
other 53 `*_tests.rs` files already live. Verified against the Codecov
API for this PR: of the 147 files it tracks, ZERO are `*_tests.rs`, so
llvm-cov emits no `SF:` record for them and they contribute no patch
lines. `src/lib.rs` keeps only the doc comment and the `mod` declaration,
neither of which is executable, so its patch contribution is now empty.

Deliberately NOT the other routes: reformatting the guard onto one line
is the anti-pattern #344's own comment warns against, and adding a
`src/**/*_tests.rs` entry to codecov.yml would change the patch
denominator for every future PR — too broad to decide here. Relocating a
test module is neither; codecov.yml's stated policy is to scope ignores
by what code IS, and this is test scaffolding.

Canary re-verified after the move: passes armed on the dev profile, and
still FAILS under `CARGO_PROFILE_DEV_DEBUG_ASSERTIONS=false` + armed.
Group green at 4036 / 4367 tests; preflight green; guard test 9 passed.
