+++ docs/development/sdlc.qmd
@@ -293,7 +293,7 @@ runs a test suite rather than a compile, so it is minutes rather than seconds.
 Name it when you want it.
 
 Its commands carry `env FERX_REQUIRE_DEBUG_ASSERTIONS=1`, which arms a canary
-(`debug_assertion_canary` in `src/lib.rs`) that fails the run if `debug_assert!`
+(`debug_assertion_canary` in `src/lib_tests.rs`) that fails the run if `debug_assert!`
 turns out to compile to nothing after all. Without it the gate would rest on the
 *absence* of a `--profile` flag, and a `[profile.dev] debug-assertions = false`
 or a `CARGO_PROFILE_DEV_DEBUG_ASSERTIONS` in the environment would neuter the
